create: allow creating sites from existing buckets
* Updated the SiteManager.create_site function to work idempotently on a
given bucket. It now skips any operations that have already been
applied. This allows safely re-running the create command on a given
bucket without breaking the site. This is useful, for example, if you
need to add an additional CNAME to your site's CloudFront distribution
(if it has one):

    $ s3site create mysite -c additional.cname.com

* Use S3 file metadata for s3site accounting. The metadata file,
__s3site.cfg, is an empty file only readable/writable by the site owner
that represents a site/bucket managed by s3site. This file's metadata is
used to store site settings. For example, the CF distribution associated
with the site is written in the metadata. Other site settings will be
stored in this file's metadata as needed in the future.

* Change --cloudfront-cname option to --cname