awsutils: add sync_bucket callbacks
files_filter - receives a mapping of local files -> S3 paths that are
scheduled to be uploaded. Callback can either modify the mapping in
place or return a new mapping of local files -> S3 paths

pre_upload_cb - called before uploading each file and receives the boto
S3 'key' which will later be used by s3site to upload the file to S3.
Callback can use this key to update the files headers before uploading
to S3. Callback can update the key in place or return a new key that
represents the file.